# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Mopro Shop API
  version: "1.0.0"
  description: |
    Mobile-first marketplace API for Mopro Shop. Buyer app + seller panel.

    ## Money format
    All monetary values are **integer minor units** (kuruş for TRY; coin-minor for TRY_COIN).
    Every monetary field is accompanied by a `currency` string field.
    Never use decimal values. Display layer divides by 100.

    ## Timestamps
    All timestamps are **RFC 3339 strings** (e.g. `"2026-01-15T10:00:00Z"`).
    Dates (no time component) use ISO 8601 `YYYY-MM-DD`.

    ## Locale resolution
    Product `title` and `description` are resolved server-side from the
    `Accept-Language` request header. The response always contains a single
    resolved string — never a `translations[]` array.
    Fallback chain: requested locale → `tr-TR` → first available locale.

    ## Step-up authentication flow
    Endpoints annotated with `stepUpAuth` require a short-lived step-up token
    (TTL 5 min, claim `scope=high_sensitivity`). Sending a regular access token
    to a `stepUpAuth` endpoint returns `403` with `code: step_up_required`.

    Client flow:
    1. Receive `403 step_up_required`.
    2. Present OTP entry screen.
    3. `POST /auth/step-up` with access token + fresh OTP code → `step_up_token`.
    4. Retry the original request with `Authorization: Bearer <step_up_token>`.

    ## X-Trace-Id
    Clients should generate a UUID per request and attach it as `X-Trace-Id`.
...

    ## Idempotency
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /addresses:
    get:
      operationId: ListAddresses
      tags: [address]
      summary: List the authenticated user's delivery addresses
      parameters:
        - $ref: '#/components/parameters/TraceId'
      responses:
        '200':
          description: Address list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: CreateAddress
      tags: [address]
      summary: Add a new delivery address
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /addresses/{id}:
    put:
      operationId: UpdateAddress
      tags: [address]
      summary: Update an existing address
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '200':
          description: Updated address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: DeleteAddress
      tags: [address]
      summary: Delete an address
      parameters:
        - $ref: '#/components/parameters/TraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Address deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Catalog ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 access token. TTL 15 min.
        Claims: sub (user_id as string), iat, exp.
        Issued by POST /auth/otp/verify or POST /auth/token/refresh.
    stepUpAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT with claim `scope=high_sensitivity`. TTL 5 min.
        Issued by POST /auth/step-up after re-verifying OTP.
        Required for: DELETE /me (account deletion).
        If a regular bearerAuth token is sent instead, the server returns
        403 with code=step_up_required.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Admin / internal token. Only for POST /products, POST /orders,
        POST /orders/{id}/refund, and GET /categories/{id}/commission.
        Not exposed to mobile clients.
  parameters:
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: |
        UUIDv7 generated client-side. Server caches the response for 24 hours
        keyed on this value. Duplicate requests within that window return the
        cached response without re-executing the operation.
      schema:
        type: string
        format: uuid
        example: "01926b7f-1234-7abc-8def-0123456789ab"
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      description: |
        Client-generated trace identifier (UUID or opaque string).
        Echoed in error responses as `error.trace_id`.
        Falls back to a server-generated UUID if absent.
      schema:
        type: string
        example: "4f3a2b1c-e71a-4c3f-b99a-8c3f2a1b7d5e"
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: Request body failed schema or business-rule validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ServiceUnavailable:
      description: |
        System is in read-only mode (maintenance window, migration in progress,
        or circuit breaker open). Retry after the `Retry-After` header value.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: service_unavailable
              message: "System is temporarily in read-only mode. Please try again shortly."
              trace_id: "4f3a2b1c"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    FieldError:
      type: object
      description: A field-level validation error within an ErrorEnvelope
      required: [name, message]
      properties:
        name:
          type: string
          example: phone
        message:
          type: string
          example: "must start with +905"
    ErrorEnvelope:
      type: object
      description: Standard error response wrapper used by all error status codes
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: string
              description: Machine-readable error slug
              example: validation_error
            message:
              type: string
              description: Human-readable error message (locale from Accept-Language)
              example: "Telefon numarası geçersiz."
            trace_id:
              type: string
              description: Request trace ID. Echoes X-Trace-Id or server-generated UUID.
              example: "4f3a2b1c-e71a-4c3f-b99a-8c3f2a1b7d5e"
            fields:
              type: array
              description: Per-field validation errors. Present only for 422 responses.
              items:
                $ref: '#/components/schemas/FieldError'
...
    Address:
      type: object
      required: [id, label, name, phone, city, district, full_address, is_default]
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
          example: "Ev"
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        district:
          type: string
        neighborhood:
          type: string
          nullable: true
        full_address:
          type: string
        postal_code:
          type: string
          nullable: true
        is_default:
          type: boolean
    AddressInput:
      type: object
      required: [label, name, phone, city, district, full_address]
      properties:
        label:
          type: string
          example: "Ev"
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        district:
          type: string
        neighborhood:
          type: string
        full_address:
          type: string
        postal_code:
          type: string
        is_default:
          type: boolean
          default: false

    # ── Catalog ───────────────────────────────────────────────────────────────
...